refactor & cleanup (frontend) : chatinterface
