+++ frontend/src/types/chatUI/TypesChatInterface.ts
@@ -1,0 +1,9 @@
+import { Message } from "../reduxSlice/TypesChatSlice";
+
+export interface IChatBubbleProps {
+  msg: Message;
+}
+
+export interface IChatInterfaceProps {
+  setSettingOpen: (isOpen: boolean) => void;
+}
