+++ frontend/src/types/reduxSlice/TypesChatSlice.ts
@@ -1,0 +1,4 @@
+export type Message = {
+  content: string;
+  sender: "user" | "assistant";
+};
